requirements.txt
# Audio
sounddevice>=0.5.0
scipy>=1.3

# ASR (also used for wake word detection)
faster-whisper>=1.0.0

# VAD
silero-vad>=5.0

# LLM client
ollama>=0.5.0
httpx>=0.27.0
anthropic>=0.40.0

# TTS (neural voices)
edge-tts>=7.0.0

# Configuration
pyyaml>=6.0

# Utilities
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
